Replace long long to unsigned long and created ConvertToBase

+++ PossibilitiesGenerator.hpp
@@ -11,8 +11,8 @@ using namespace std;
 // File Starts Here
 class PossibilitiesGenerator {
    private:
-    static long long *increase(long long *vec, long long len, long long base) {
-        for (long long i = 0; i < len; i++) {
+    static unsigned long *increase(unsigned long *vec, unsigned long len, unsigned long base) {
+        for (unsigned long i = 0; i < len; i++) {
             if(vec[i] == (base - 1))
                 vec[i] = 0;
             else{
@@ -21,24 +21,24 @@ class PossibilitiesGenerator {
             }
         }
 
-        long long *new_vec = (long long *) calloc(len, sizeof(long long));
-        for (long long i = 0; i < len; i++) {
+        unsigned long *new_vec = (unsigned long *) calloc(len, sizeof(unsigned long));
+        for (unsigned long i = 0; i < len; i++) {
             new_vec[i] = vec[i];
         }
         return new_vec;
     }
 
-    static bool is_the_last(long long *vec, long long len, long long base) {
-        for (long long i = 0; i < len; i++){
+    static bool is_the_last(unsigned long *vec, unsigned long len, unsigned long base) {
+        for (unsigned long i = 0; i < len; i++){
             if (vec[i] != base - 1) return false;
         }
         return true;
     }
 
    public:
-    static vector<long long *> generate_all_possibilites(long long len, long long base) {
-        vector<long long *> possibilities;
-        long long *base_possibility = (long long *) calloc(len, sizeof(long long));
+    static vector<unsigned long *> generate_all_possibilites(unsigned long len, unsigned long base) {
+        vector<unsigned long *> possibilities;
+        unsigned long *base_possibility = (unsigned long *) calloc(len, sizeof(unsigned long));
 
         int i = 0;
         do {
